+++ src/routes/pages/Checkout.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'devextreme/dist/css/dx.light.css';
-import './App.css';
+import '../../App.css';
 
 export default function Checkout() {
     return (
