+++ src/routes/pages/Home.js
@@ -1,4 +1,4 @@
-import './App.css';
+import "../../App.css";
 import React from 'react';
 
 function Item(props) {
